Some changes

+++ API/Controllers/PostController.cs
@@ -1,12 +1,15 @@
 using System;
+using System.IO;
+using System.IO.Compression;
+using System.Linq;
 using System.Collections.Generic;
 using System.Text.RegularExpressions;
-using System.IO.Compression;
+using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Cors;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Hosting;
 using Newtonsoft.Json.Linq;
-using System.Linq;
 using API.Models;
 
 namespace API.Controllers
@@ -84,17 +87,19 @@ namespace API.Controllers
                 return NotFound();
             }
 
-            var matches = Regex.Matches(post.Content, "<img.+?src=[\"'](.+?)[\"'].+?>", RegexOptions.IgnoreCase); //.Groups[1].Value; 
-
-            foreach (Match match in matches)
+            if (post.Content != null)
             {
-                string urlPath = match.Groups[1].Value;
-                string fileName = System.IO.Path.GetFileName(urlPath);
-                string fullPath = System.IO.Path.GetFullPath("wwwroot/uploads/" + fileName);
-                if (System.IO.File.Exists(fullPath))
-                    System.IO.File.Delete(fullPath);
+                var matches = Regex.Matches(post.Content, "<img.+?src=[\"'](.+?)[\"'].+?>", RegexOptions.IgnoreCase); //.Groups[1].Value; 
+                foreach (Match match in matches)
+                {
+                    string urlPath = match.Groups[1].Value;
+                    string fileName = System.IO.Path.GetFileName(urlPath);
+                    string fullPath = System.IO.Path.GetFullPath("wwwroot/uploads/" + fileName);
+                    if (System.IO.File.Exists(fullPath))
+                        System.IO.File.Delete(fullPath);
+                }
             }
-
+            
             PostItems.Remove(id);
             return new NoContentResult();
         }
@@ -103,17 +108,18 @@ namespace API.Controllers
         [HttpGet("DownloadZip")]
         public IActionResult DownloadZip()
         {
-            string webRootPath = _hostingEnvironment.WebRootPath;
-            string uploadPath = System.IO.Path.Combine(webRootPath, "uploads");
-            string exportPath = System.IO.Path.Combine(webRootPath, "exports");
-            string jsonFile = System.IO.Path.Combine(uploadPath, "posts.json");
-            string zipName = string.Format("export-{0}.zip", DateTime.Now.ToString("yyyy-MM-dd"));
-            string zipFile = System.IO.Path.Combine(exportPath, zipName);
-            
-            var posts = PostItems.GetAll();
+            string webrootPath = _hostingEnvironment.WebRootPath;
+            string uploadPath = System.IO.Path.Combine(webrootPath, "uploads");
+            string jsonFileName = "export.json";
+            string jsonFilePath = System.IO.Path.Combine(webrootPath + @"\", jsonFileName);
+            string zipFileName = string.Format("export-{0}.zip", DateTime.Now.ToString("yyyy-MM-dd"));
+            string zipFilePath = System.IO.Path.Combine(webrootPath + @"\", zipFileName);
+
+            DirectoryInfo di = new System.IO.DirectoryInfo(webrootPath);
+            foreach (var file in di.EnumerateFiles("export*.*")) { file.Delete(); }
 
             Newtonsoft.Json.Linq.JArray json = new JArray(
-                posts.Select(p => new JObject
+                PostItems.GetAll().Select(p => new JObject
                 {
                     { "Title", p.Title},
                     { "Creation", p.Creation},
@@ -121,24 +127,86 @@ namespace API.Controllers
                 })
             );
 
-            if (System.IO.File.Exists(jsonFile))
-                System.IO.File.Delete(jsonFile);
+            System.IO.File.WriteAllText(jsonFilePath, json.ToString());
 
-            System.IO.File.WriteAllText(jsonFile, json.ToString());
+            ZipFile.CreateFromDirectory(uploadPath, zipFilePath);
 
-            System.IO.DirectoryInfo di = new System.IO.DirectoryInfo(exportPath);
+            using (ZipArchive archive = ZipFile.Open(zipFilePath, ZipArchiveMode.Update))
+            {
+                archive.CreateEntryFromFile(jsonFilePath, jsonFileName);
+            }
 
-            if (System.IO.File.Exists(zipFile))
-                System.IO.File.Delete(zipFile);
+            var fileStream = new System.IO.FileStream(zipFilePath, System.IO.FileMode.Open);
+            var memoryStream = new System.IO.MemoryStream();
+            fileStream.CopyTo(memoryStream);
+            fileStream.Close();
+            memoryStream.Position = 0;
 
-            ZipFile.CreateFromDirectory(uploadPath, zipFile);
+            foreach (var file in di.EnumerateFiles("export*.*")) { file.Delete(); }
 
-            if (System.IO.File.Exists(jsonFile))
-                System.IO.File.Delete(jsonFile);
+            return File(memoryStream, "application/octetstream", zipFileName);
+        }
 
-            var stream = new System.IO.FileStream(zipFile, System.IO.FileMode.Open);
+        [Route("api/[controller]/UploadZip")]
+        [HttpPost("UploadZip")]
+        public async Task<IActionResult> UploadZip(IFormFile file)
+        {
+            long size = file.Length;
+
+            string webrootPath = _hostingEnvironment.WebRootPath;
+            string appRoot = Environment.CurrentDirectory;
+            string uploadPath = System.IO.Path.Combine(webrootPath, "uploads");
+            string jsonFilePath = System.IO.Path.Combine(uploadPath + @"\", @"export.json");
+            string importFilePath = System.IO.Path.Combine(appRoot + @"\", @"App_Data\import.zip");
+
+            if (file.Length == 0) throw new Exception("Le fichier est vide");
+            if (file.FileName.EndsWith(".zip") == false) throw new Exception("Le type du fichier n'est pas valide");
+
+            using (var stream = new System.IO.FileStream(importFilePath, System.IO.FileMode.Create))
+            {
+                await file.CopyToAsync(stream);
+            }
+
+            using (ZipArchive archive = ZipFile.OpenRead(importFilePath))
+            {
+                foreach (ZipArchiveEntry entry in archive.Entries)
+                {
+                    string fileName = Path.Combine(uploadPath, entry.Name);
+                    if (!System.IO.File.Exists(fileName))
+                    {
+
+                        entry.ExtractToFile(fileName);
+                    }
+                }
+            }
+
+            string json = System.IO.File.ReadAllText(jsonFilePath);
+
+            var elements = Newtonsoft.Json.JsonConvert.DeserializeObject<JArray>(json).ToObject<List<JObject>>();
+
+            foreach(var elm in elements) 
+            {
+                Console.WriteLine("t = " + elm["Title"]);
+                Console.WriteLine("l = " + elm["Content"]);
+                Console.WriteLine("c = " + elm["Creation"]);
+
+                PostItem postItem = new PostItem
+                {
+                    Title = (string)elm["Title"],
+                    Content = (string)elm["Content"],
+                    Creation = (DateTime)elm["Creation"]
+                };
+
+                PostItems.Add(postItem);
+            }
             
-            return File(stream, "application/octetstream", zipName);
+            if (System.IO.File.Exists(jsonFilePath))
+                System.IO.File.Delete(jsonFilePath);
+
+            if (System.IO.File.Exists(importFilePath))
+                System.IO.File.Delete(importFilePath);
+
+            return Ok(new { count = elements.Count });
         }
     }
 }
